Migrate code to common
