+++ sematic/ui/packages/main/src/components/Shell.tsx
@@ -2,7 +2,7 @@ import Box from "@mui/material/Box";
 import { ThemeProvider } from "@mui/material/styles";
 import { useContext } from "react";
 import { Navigate, Outlet } from "react-router-dom";
-import { UserContext } from "../appContext";
+import UserContext from "@sematic/common/src/context/UserContext";
 import { useAppContext } from "../hooks/appHooks";
 import createTheme from "@sematic/common/src/theme/mira";
 import SideBar from "./SideBar";
