+++ sematic/ui/packages/main/src/hooks/metricsHooks.ts
@@ -1,4 +1,4 @@
-import { useHttpClient } from "src/hooks/httpHooks";
+import { useHttpClient } from "@sematic/common/src/hooks/httpHooks";
 import useAsync from "react-use/lib/useAsync";
 import { BasicMetricsPayload } from "src/Payloads";
