+++ sematic/ui/packages/main/src/hooks/podHistoryHooks.ts
@@ -1,10 +1,10 @@
 import { useRefFn } from "@sematic/common/lib/src/utils/hooks";
 import { Job } from "@sematic/common/src/Models";
+import { useHttpClient } from "@sematic/common/src/hooks/httpHooks";
 import { useCallback, useEffect } from "react";
 import useAsyncFn from "react-use/lib/useAsyncFn";
 import useLatest from "react-use/lib/useLatest";
 import { RunJobPayload } from "src/Payloads";
-import { useHttpClient } from "src/hooks/httpHooks";
 import { jobSocket } from "src/sockets";
 import { AsyncInvocationQueue } from "src/utils";
