+++ sematic/ui/tsconfig.base.json
@@ -20,6 +20,7 @@
     "isolatedModules": true,
     "jsx": "react-jsx",
     "types": [
+      "node",
       "cypress",
     ],
   }
